+++ src/QuickMarcEditor/QuickMarcDeriveWrapper.js
@@ -36,6 +36,7 @@ const propTypes = {
   initialValues: PropTypes.object.isRequired,
   instance: PropTypes.object,
   marcType: PropTypes.oneOf(Object.values(MARC_TYPES)).isRequired,
+  marcSpec: PropTypes.object.isRequired,
   mutator: PropTypes.object.isRequired,
   onClose: PropTypes.func.isRequired,
 };
@@ -47,6 +48,7 @@ const QuickMarcDeriveWrapper = ({
   initialValues,
   mutator,
   marcType,
+  marcSpec,
 }) => {
   const showCallout = useShowCallout();
   const { linkableBibFields, actualizeLinks, linkingRules } = useAuthorityLinking({ marcType, action });
@@ -62,11 +64,11 @@ const QuickMarcDeriveWrapper = ({
       autopopulatePhysDescriptionField,
       autopopulateMaterialCharsField,
       marcRecord => autopopulateSubfieldSection(marcRecord, marcType),
-      marcRecord => cleanBytesFields(marcRecord, marcType),
+      marcRecord => cleanBytesFields(marcRecord, marcSpec),
     )(formValues);
 
     return formValuesForDerive;
-  }, [marcType]);
+  }, [marcType, marcSpec]);
 
   const validate = useCallback((formValues) => {
     const formValuesForValidation = prepareForSubmit(formValues);
@@ -157,6 +159,7 @@ const QuickMarcDeriveWrapper = ({
       onSubmit={onSubmit}
       action={action}
       marcType={marcType}
+      marcSpec={marcSpec}
       httpError={httpError}
       confirmRemoveAuthorityLinking
       validate={validate}