+++ src/QuickMarcEditor/QuickMarcEditorRows/QuickMarcEditorRows.js
@@ -79,6 +79,7 @@ const QuickMarcEditorRows = ({
     updateRecord,
   },
   marcType,
+  marcSpec,
   instance,
   linksCount,
   isLoadingLinkSuggestions,
@@ -292,7 +293,7 @@ const QuickMarcEditorRows = ({
               isRequestToCentralTenantFromMember,
               onCheckCentralTenantPerm,
             );
-            const canBeLinkedAuto = isRecordForAutoLinking(recordRow, autoLinkableBibFields);
+            const canBeLinkedAuto = isRecordForAutoLinking(recordRow, autoLinkableBibFields) || false;
 
             const canViewAuthorityRecord = stripes.hasPerm('ui-marc-authorities.authority-record.view') && recordRow._isLinked;
 
@@ -469,7 +470,7 @@ const QuickMarcEditorRows = ({
                   {
                     isFixedField && (
                       FixedFieldFactory.getFixedField(
-                        `${name}.content`, marcType, type, subtype,
+                        `${name}.content`, marcType, marcSpec, type, subtype,
                       )
                     )
                   }
@@ -594,6 +595,7 @@ QuickMarcEditorRows.propTypes = {
     updateRecord: PropTypes.func.isRequired,
   }),
   marcType: PropTypes.oneOf(Object.values(MARC_TYPES)).isRequired,
+  marcSpec: PropTypes.object.isRequired,
   onCheckCentralTenantPerm: PropTypes.func,
 };