+++ src/common/constants/api.js
@@ -9,3 +9,4 @@ export const EXTERNAL_INSTANCE_APIS = {
 export const MARC_RECORD_API = 'records-editor/records';
 export const MARC_RECORD_STATUS_API = `${MARC_RECORD_API}/status`;
 export const LINKING_RULES_API = 'linking-rules/instance-authority';
+export const MARC_SPEC_API = 'marc-specifications';